pyproject.toml
[project]
name = "college-content-generator"
version = "1.0.0"
description = "AI-powered content generation system for college and education content"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ai", "content-generation", "education", "llm", "streamlit"]

dependencies = [
    "streamlit>=1.31.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "google-search-results>=2.4.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "database", "models", "agents", "utils"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
